Gibbs Sampler for Multivariate Regression with Covariance Option

This function performs a single Gibbs sampling draw for a multivariate regression model with an option to use either a dense or diagonal covariance matrix for the outcomes. Each outcome has its own set of predictors.

gibbs_sampler_one_draw_cov(
  y_list,
  X_list,
  sigma_param,
  covariance_type = "diagonal"
)

Arguments

y_list

A list of outcome vectors. Each element \(y_j\) in the list corresponds to the outcomes for one dependent variable.

X_list

A list of predictor matrices. Each element \(X_j\) in the list corresponds to the predictor matrix for the corresponding outcome \(y_j\).

sigma_param

A list of initial values for the residual variances \(\sigma^2_j\) for each outcome when covariance_type = "diagonal", or a dense covariance matrix when covariance_type = "dense".

covariance_type

A string specifying the type of covariance matrix to use. Can be either "diagonal" (no correlation between outcomes) or "dense" (allows correlation between outcomes).

Value

A list with the following elements:

beta_samples

A list of sampled regression coefficients \(\beta_j\) for each outcome.

sigma2_samples

A list of sampled residual variances \(\sigma^2_j\) for each outcome (diagonal case), or the sampled covariance matrix \(\Sigma\) (dense case).

residuals

A list of residuals for each outcome \(r_j = y_j - X_j \beta_j\).

Details

The model assumes that each outcome \(y_j\) follows a regression model: $$y_j = X_j \beta_j + \epsilon_j, \quad \epsilon_j \sim \mathcal{N}(0, \Sigma)$$ where \(X_j\) is the predictor matrix for outcome \(y_j\), \(\beta_j\) is the vector of regression coefficients, and \(\Sigma\) is the covariance matrix of the residuals. The covariance matrix can either be diagonal (independent outcomes) or dense (allowing correlation between outcomes).

Examples

# Simulate data for two outcomes with different predictor matrices
set.seed(123)
n <- 100  # Number of observations
p <- 2    # Number of outcomes

# Create predictor matrices X for each outcome
X1 <- cbind(1, rnorm(n))
X2 <- cbind(1, rnorm(n), rnorm(n))

# True beta values
beta1 <- c(1, 0.5)
beta2 <- c(1, -0.3, 0.2)

# Simulate outcomes
y1 <- X1 %*% beta1 + rnorm(n)
y2 <- X2 %*% beta2 + rnorm(n)

# Prepare data for the Gibbs sampler
y_list <- list(y1, y2)
X_list <- list(X1, X2)
sigma_param <- list(1, 1)  # Diagonal case: Initial values for sigma^2

# Run the Gibbs sampler with diagonal covariance
result_diag <-
    gibbs_sampler_one_draw_cov(y_list, X_list,
                              sigma_param,
                              covariance_type = "diagonal")
print(result_diag)
#> $beta_samples
#> $beta_samples[[1]]
#> [1] 1.0846891 0.4270266
#> 
#> $beta_samples[[2]]
#> [1]  1.0247100 -0.5886832  0.1286928
#> 
#> 
#> $sigma2_samples
#> $sigma2_samples[[1]]
#> [1] 0.9473082
#> 
#> $sigma2_samples[[2]]
#> [1] 1.129763
#> 
#> 
#> $residuals
#>              V1          V2
#> 1   -0.84083109 -0.14655748
#> 2   -0.85417488 -1.02561895
#> 3   -0.90948351 -0.74958075
#> 4   -1.13205712 -0.11514758
#> 5   -0.51241404  0.34172436
#> 6    0.37164425 -1.72221505
#> 7   -2.06526494 -0.65728591
#> 8    0.03497552  0.20778928
#> 9    1.10186397 -0.55556260
#> 10   1.92036347  0.46402822
#> 11   1.30581235  0.30992868
#> 12   0.69834253  0.43601078
#> 13  -1.78217381  0.24937658
#> 14  -0.67811889 -0.20025746
#> 15  -0.47729716 -0.35921521
#> 16   0.74923199 -2.46143183
#> 17  -0.15403056 -0.11860212
#> 18  -1.48684848  0.16905394
#> 19   1.65092697  0.17723844
#> 20   0.79220101 -0.96724645
#> 21   0.07481845  1.74782714
#> 22   1.11751316  0.03226131
#> 23  -1.49833441  0.03913034
#> 24   0.52294154  1.22408343
#> 25  -0.65321270 -0.23681563
#> 26   0.47597267 -0.65899493
#> 27  -0.08437485  2.39045190
#> 28   0.55946380 -0.04221756
#> 29   1.16777480  1.39426857
#> 30   0.01409617 -1.55622105
#> 31   0.96399011  0.34121719
#> 32  -1.29465549  0.47761513
#> 33  -0.74097314  0.30253000
#> 34   1.49860904 -1.20497624
#> 35   0.35265248 -0.63914659
#> 36  -2.08665946 -0.86944527
#> 37  -1.40830526  0.25328730
#> 38  -0.28998800  1.30355196
#> 39   0.75876319 -1.67544638
#> 40  -0.21433660  0.73842673
#> 41   0.48880325 -1.11939552
#> 42   0.85914387  0.31855696
#> 43   1.49402546  0.28802097
#> 44   0.12960379 -0.74295670
#> 45  -0.04852187 -1.14492809
#> 46  -1.91988351  0.12299930
#> 47  -0.01476136 -0.01507570
#> 48  -0.69059257  0.96021649
#> 49  -1.00178194 -1.51818367
#> 50  -0.27067903 -1.99801647
#> 51   0.94873961  1.60631616
#> 52  -2.07952072  1.20386264
#> 53  -0.51509677  0.48195653
#> 54   0.13181979  0.40582017
#> 55  -0.99437193  0.97196612
#> 56   0.35987591 -2.77289349
#> 57   0.21372306  1.32516147
#> 58  -0.07506397 -0.57223707
#> 59  -2.54154920  0.47983930
#> 60   2.50252706 -0.53730981
#> 61  -0.26228474  1.11399568
#> 62   0.52984795 -0.71099340
#> 63   0.16476213  0.13337067
#> 64   0.86565523  0.61295217
#> 65   0.65475810 -1.07432400
#> 66  -0.27233272  1.38174744
#> 67   0.32618610  0.89046239
#> 68  -1.02623001  1.43464001
#> 69   0.83953494  0.16193056
#> 70  -0.39612573  1.21316775
#> 71   2.29625206  1.94878785
#> 72  -1.90424592 -0.21870794
#> 73  -0.47528414 -1.30869955
#> 74   0.68893798  0.25089366
#> 75   0.37523713 -0.39217681
#> 76  -0.59933067 -0.22126783
#> 77  -1.10225068  1.71096858
#> 78  -0.02929331 -0.25590725
#> 79  -0.08576615  0.50484303
#> 80  -1.88510569 -0.38295687
#> 81  -0.04971738 -0.43012649
#> 82   0.13365647  0.70652543
#> 83   0.06298899  1.23009302
#> 84  -1.09268376 -0.17219342
#> 85   0.37535454  0.62834699
#> 86   1.31809233  0.70678663
#> 87   0.45158742  1.22623888
#> 88  -1.18852087 -0.45763821
#> 89  -0.54411889  1.80423403
#> 90   0.34524696 -0.53781962
#> 91  -0.65923534  0.01472675
#> 92  -2.20231701  1.36077961
#> 93   0.81698281  1.37834408
#> 94  -0.95998714 -1.41435574
#> 95  -0.55895789 -1.13339594
#> 96   1.37540854 -0.80147090
#> 97  -0.69921684  0.46367165
#> 98   0.87288230 -0.31741907
#> 99  -1.36257182  0.16446771
#> 100 -0.51413292  0.27434748
#> 
#> $predicted
#>             V1          V2
#> 1   0.84535108  1.72588550
#> 2   0.98639717  1.04238500
#> 3   1.75029896  1.13581103
#> 4   1.11479803  1.29920762
#> 5   1.13989838  1.53158919
#> 6   1.81706742  0.98992744
#> 7   1.28151255  1.38528245
#> 8   0.54447430  1.93007629
#> 9   0.79138465  1.46100289
#> 10  0.89437957  0.47675909
#> 11  1.60740454  1.37875303
#> 12  1.23833914  0.69817242
#> 13  1.25582914  2.13574108
#> 14  1.13195354  0.99100462
#> 15  0.84733014  0.59121511
#> 16  1.84774849  1.06307625
#> 17  1.29728446  0.90572560
#> 18  0.24489127  1.30882937
#> 19  1.38418669  1.36581708
#> 20  0.88279458  1.46226359
#> 21  0.62869997  0.88158710
#> 22  0.99160799  1.66200265
#> 23  0.64655790  1.45632238
#> 24  0.77343314  1.26652867
#> 25  0.81778069 -0.10754065
#> 26  0.36442620  1.41619127
#> 27  1.44244641  0.79546531
#> 28  1.15018347  0.88651486
#> 29  0.59867435  1.70478695
#> 30  1.62010138  0.93598832
#> 31  1.26680063  0.42595954
#> 32  0.95868571  0.74729363
#> 33  1.46693153  1.02804644
#> 34  1.45967542  1.17838349
#> 35  1.43552604  2.15950229
#> 36  1.37875677  0.18922023
#> 37  1.32122664  1.86102297
#> 38  1.05825113  0.64303535
#> 39  0.95403488 -0.05741152
#> 40  0.92221784  1.77412738
#> 41  0.78803073  0.51009132
#> 42  0.99590287  1.11443181
#> 43  0.54433120  2.14273705
#> 44  2.01089093  1.77000644
#> 45  1.60052096  1.94446466
#> 46  0.60509186  1.58206604
#> 47  0.91264654  1.87222610
#> 48  0.88541483  0.44474316
#> 49  1.41775491 -0.29713988
#> 50  1.04908827  1.84483834
#> 51  1.19286282  0.51264396
#> 52  1.07249885  0.49967827
#> 53  1.06638225  0.78488820
#> 54  1.66911863  1.62997055
#> 55  0.98827886  1.30074634
#> 56  1.73226234  1.17837627
#> 57  0.42333046  0.83237866
#> 58  1.33433469  1.32513192
#> 59  1.13757813  0.43495753
#> 60  1.17690187  1.04794587
#> 61  1.24680523  0.33793237
#> 62  0.87018361  1.57929996
#> 63  0.94240067  1.77261056
#> 64  0.64973031 -0.71590940
#> 65  0.62700573  1.56520986
#> 66  1.21430387  1.04831101
#> 67  1.27608657  0.63283653
#> 68  1.10732329  1.08345094
#> 69  1.47852180  0.67040878
#> 70  1.96012973  0.81898708
#> 71  0.87500571  1.26024796
#> 72  0.09861258  1.11014351
#> 73  1.51416616  1.13283050
#> 74  0.78184150 -0.40783566
#> 75  0.79089111  1.57046501
#> 76  1.52263531  1.61243577
#> 77  0.96308343  1.02496041
#> 78  0.56341016  0.85152937
#> 79  1.16211048  0.82283793
#> 80  1.02537877  1.29771740
#> 81  1.08715054  1.43608002
#> 82  1.24921405  0.37587588
#> 83  0.92640739  1.28022210
#> 84  1.35985499  1.50003533
#> 85  0.99053545  1.17900811
#> 86  1.22636879  1.15803389
#> 87  1.55306849  0.39976220
#> 88  1.27052314  1.18599170
#> 89  0.94550762  0.55262097
#> 90  1.57526047  1.34026357
#> 91  1.50894163  1.04883226
#> 92  1.31886916  1.35151264
#> 93  1.18663387  1.11641739
#> 94  0.81655649  1.47747928
#> 95  1.66572384  2.05406195
#> 96  0.82836228 -0.14243200
#> 97  2.01873841  0.91133327
#> 98  1.73915455  1.58746081
#> 99  0.98403876  1.38719347
#> 100 0.64638007  1.88343718
#> 
#> $covariance
#> NULL
#> 

# Simulate a dense covariance matrix
sigma_param_dense <- matrix(c(1, 0.5, 0.5, 1), 2, 2)  # Dense covariance matrix

# Run the Gibbs sampler with dense covariance
result_dense <-
    gibbs_sampler_one_draw_cov(y_list, X_list,
                              sigma_param_dense,
                              covariance_type = "dense")
print(result_dense)
#> $beta_samples
#> $beta_samples[[1]]
#> [1] 0.8916744 0.5225635
#> 
#> $beta_samples[[2]]
#> [1]  1.1237785 -0.3095355  0.1796746
#> 
#> 
#> $sigma2_samples
#>             [,1]        [,2]
#> [1,]  1.12823193 -0.03095692
#> [2,] -0.03095692  1.04562036
#> 
#> $residuals
#>               V1           V2
#> 1   -0.594270239 -0.159416786
#> 2   -0.639169717 -1.263305042
#> 3   -0.865383054 -0.766268235
#> 4   -0.945778567 -0.144893347
#> 5   -0.331751091  0.529421755
#> 6    0.400806862 -1.784434289
#> 7   -1.916284769 -0.497045810
#> 8    0.348850329  0.604637546
#> 9    1.360498512 -0.632657868
#> 10   2.155955363  0.111178357
#> 11   1.381882001  0.365387620
#> 12   0.856981721  0.154806812
#> 13  -1.627447591  0.539102563
#> 14  -0.495678471 -0.257506151
#> 15  -0.231179080 -0.552675325
#> 16   0.771530443 -2.729996539
#> 17  -0.008578973 -0.224678644
#> 18  -1.105949146  0.285700188
#> 19   1.776936266  0.378390287
#> 20   1.030384773 -0.714934737
#> 21   0.369849790  1.645180003
#> 22   1.331352528  0.166172177
#> 23  -1.207298360  0.020417836
#> 24   0.785592298  1.160428389
#> 25  -0.400483644 -0.832054120
#> 26   0.830128931 -0.579119285
#> 27   0.028600233  2.261597310
#> 28   0.737825704 -0.126483650
#> 29   1.469523648  1.518599112
#> 30   0.087325215 -1.583607414
#> 31   1.116261725 -0.260778651
#> 32  -1.073450545  0.257114948
#> 33  -0.633476015  0.181013845
#> 34   1.607729540 -1.148454312
#> 35   0.467175824 -0.135772523
#> 36  -1.959435353 -1.217180218
#> 37  -1.268210165  0.571278960
#> 38  -0.091058436  0.976568350
#> 39   0.981008638 -2.323970377
#> 40   0.015027150  1.082261764
#> 41   0.748188154 -1.374160208
#> 42   1.072022360  0.318283230
#> 43   1.807932285  0.551541131
#> 44   0.115403034 -0.361227510
#> 45   0.029087823 -0.787803093
#> 46  -1.619570419  0.075146427
#> 47   0.216743735  0.299049377
#> 48  -0.452995028  0.738444614
#> 49  -0.883282728 -2.169601684
#> 50  -0.069699498 -1.762562973
#> 51   1.117553041  1.306501033
#> 52  -1.883778741  0.918763171
#> 53  -0.317986346  0.307687920
#> 54   0.194082399  0.583623997
#> 55  -0.779787745  0.824747781
#> 56   0.408011624 -2.789175085
#> 57   0.554700903  1.013834694
#> 58   0.062098515 -0.599497092
#> 59  -2.360367153  0.113844467
#> 60   2.674911371 -0.453664945
#> 61  -0.105539638  0.747632637
#> 62   0.770853110 -0.492212116
#> 63   0.389610462  0.383667584
#> 64   1.155981536 -0.457131581
#> 65   0.950168485 -1.173932805
#> 66  -0.108316221  1.120531015
#> 67   0.476380203  0.620485174
#> 68  -0.838279163  1.560168546
#> 69   0.944439021 -0.061598136
#> 70  -0.398969881  1.006555696
#> 71   2.536178388  1.866762082
#> 72  -1.490620233 -0.385074197
#> 73  -0.378354636 -1.433145549
#> 74   0.949707576 -0.371193824
#> 75   0.633982090 -0.327619328
#> 76  -0.504295931  0.008374755
#> 77  -0.882029623  1.592439778
#> 78   0.280345053 -0.445446502
#> 79   0.089927370  0.262091307
#> 80  -1.678821724 -0.355331723
#> 81   0.142746638 -0.147359187
#> 82   0.289862656  0.217293828
#> 83   0.291415432  1.208948307
#> 84  -0.961230829 -0.016112627
#> 85   0.589433868  0.589212158
#> 86   1.479409592  0.655925793
#> 87   0.539813460  0.806070708
#> 88  -1.037082075 -0.664014121
#> 89  -0.319965667  1.505840699
#> 90   0.428508081 -0.506080181
#> 91  -0.561136967 -0.203769907
#> 92  -2.061694482  1.298602448
#> 93   0.987189818  1.194485291
#> 94  -0.706984194 -1.234045314
#> 95  -0.495935779 -0.968647246
#> 96   1.625770224 -1.461457472
#> 97  -0.715173280  0.101741306
#> 98   0.919476046  0.001673297
#> 99  -1.147039012  0.234934978
#> 100 -0.223057079  0.442480200
#> 
#> $predicted
#>              V1        V2
#> 1    0.59879023 1.7387448
#> 2    0.77139200 1.2800711
#> 3    1.70619851 1.1524985
#> 4    0.92851948 1.3289534
#> 5    0.95923543 1.3438918
#> 6    1.78790480 1.0521467
#> 7    1.13253237 1.2250423
#> 8    0.23059949 1.5332280
#> 9    0.53275011 1.5380982
#> 10   0.65878767 0.8296090
#> 11   1.53133489 1.3232941
#> 12   1.07969996 0.9793764
#> 13   1.10110292 1.8460151
#> 14   0.94951312 1.0482533
#> 15   0.60121206 0.7846752
#> 16   1.82545003 1.3316410
#> 17   1.15183288 1.0118021
#> 18  -0.13600806 1.1921831
#> 19   1.25817739 1.1646652
#> 20   0.64461082 1.2099519
#> 21   0.33366863 0.9842342
#> 22   0.77776862 1.5280918
#> 23   0.35552185 1.4750349
#> 24   0.51078239 1.3301837
#> 25   0.56505163 0.4876978
#> 26   0.01026994 1.3363156
#> 27   1.32947133 0.9243199
#> 28   0.97182157 0.9707809
#> 29   0.29692550 1.5804564
#> 30   1.54687234 0.9633747
#> 31   1.11452902 1.0279554
#> 32   0.73748077 0.9677938
#> 33   1.35943441 1.1495626
#> 34   1.35055491 1.1218616
#> 35   1.32100269 1.6561282
#> 36   1.25153266 0.5369552
#> 37   1.18113154 1.5430313
#> 38   0.85932157 0.9700190
#> 39   0.73178943 0.5911125
#> 40   0.69285409 1.4302923
#> 41   0.52864583 0.7648560
#> 42   0.78302438 1.1147055
#> 43   0.23042437 1.8792169
#> 44   2.02509168 1.3882773
#> 45   1.52291127 1.5873397
#> 46   0.30477877 1.6299189
#> 47   0.68114144 1.5581010
#> 48   0.64781729 0.6665150
#> 49   1.29925570 0.3542781
#> 50   0.84810873 1.6093848
#> 51   1.02404939 0.8124591
#> 52   0.87675687 0.7847777
#> 53   0.86927183 0.9591568
#> 54   1.60685603 1.4521667
#> 55   0.77369468 1.4479647
#> 56   1.68412662 1.1946579
#> 57   0.08235262 1.1437054
#> 58   1.19717220 1.3523919
#> 59   0.95639608 0.8009524
#> 60   1.00451756 0.9643010
#> 61   1.09006012 0.7042954
#> 62   0.62917845 1.3605187
#> 63   0.71755234 1.5223136
#> 64   0.35940401 0.3541744
#> 65   0.33159535 1.6648187
#> 66   1.05028737 1.3095274
#> 67   1.12589246 0.9028137
#> 68   0.91937244 0.9579224
#> 69   1.37361772 0.8939375
#> 70   1.96297389 1.0255991
#> 71   0.63507938 1.3422737
#> 72  -0.31501310 1.2765098
#> 73   1.41723666 1.2572765
#> 74   0.52107191 0.2142518
#> 75   0.53214615 1.5059075
#> 76   1.42760058 1.3827932
#> 77   0.74286238 1.1434892
#> 78   0.25377180 1.0410686
#> 79   0.98641696 1.0655897
#> 80   0.81909481 1.2700923
#> 81   0.89468652 1.1533127
#> 82   1.09300786 0.8651075
#> 83   0.69798095 1.3013668
#> 84   1.22840206 1.3439545
#> 85   0.77645613 1.2181429
#> 86   1.06505153 1.2088947
#> 87   1.46484245 0.8199304
#> 88   1.11908435 1.3923676
#> 89   0.72135440 0.8510143
#> 90   1.49199935 1.3085241
#> 91   1.41084326 1.2673289
#> 92   1.17824663 1.4136898
#> 93   1.01642687 1.3002762
#> 94   0.56355354 1.2971688
#> 95   1.60270173 1.8893133
#> 96   0.57800059 0.5175546
#> 97   2.03469485 1.2732636
#> 98   1.69256081 1.2683684
#> 99   0.76850595 1.3167262
#> 100  0.35530423 1.7153045
#> 
#> $covariance
#>             [,1]        [,2]
#> [1,]  1.12823193 -0.03095692
#> [2,] -0.03095692  1.04562036
#>